requirements.txt
ffmpeg_python==0.2.0
librosa==0.9.1
matplotlib==3.5.2
numpy==1.22.4
pandas==1.4.2
scikit_learn==1.1.1
scipy==1.8.1
sorcery==0.2.2
xgboost==1.6.1
confuse==1.7.0
PyYAML==6.0
scikit-learn==1.1.1
requests==2.28.0
optuna==2.10.1
utm==0.7.0